more robust to prevent pixels with no data on PC with low resources

+++ hostApps/python/modules/h5Reader.py
@@ -155,6 +155,7 @@ class h5Reader:
         init of the h5reader class
         """
         # settings for the measurements
+        self.hfFileOpenSuccess = False
         self.h5 = None
         self.deleteAfter = deleteAfter
         self.scripts_path = os.path.dirname(__file__)
@@ -269,6 +270,7 @@ class h5Reader:
     def tryOpen(self, nRetry=0):
         try:
             self.h5 = h5py.File(self.hfFile, 'r+')
+            self.hfFileOpenSuccess = True
         except BlockingIOError as ex:
             """
             [Errno 11] Unable to open file
@@ -276,16 +278,22 @@ class h5Reader:
             error message = 'Resource temporarily
             unavailable')
             """
-            if nRetry < 10:
+            if nRetry < 20:
                 # nRetry is increased on each recursion
                 dbgPrint(f"File busy")
-                timeSleep(0.001)
+                if nRetry < 10:
+                    # normal operation
+                    timeSleep(0.001)
+                else:
+                    # slow system with limited resources
+                    timeSleep(0.010)
                 self.tryOpen(nRetry=nRetry+1)
             else:
                 # max number of retries reached
                 # consider reading the file as failed
                 dbgPrint(f"File still busy")
                 self.h5 = None # make sure it is still set to None
+                self.hfFileOpenSuccess = False
 
     def h5Open(self):
         """
@@ -304,7 +312,8 @@ class h5Reader:
             dbgPrint(f"Closing file {self.hfFile}")
             self.h5.close()
             self.h5 = None
-        if self.deleteAfter and os.path.isfile(self.hfFile):
+        if self.deleteAfter and os.path.isfile(self.hfFile) and self.hfFileOpenSuccess:
+            # NOTE: delete only file successfully parsed
             dbgPrint(f"Deleting file {self.hfFile}")
             os.remove(self.hfFile)
             self.hfFile = ""
